+++ nlp/fasttext/__init__.py
@@ -1,0 +1,1 @@
+from fasttext.main import *
